requirements.txt
Django>=1.9
django-photologue>=3.5.1
bpython>=0.12
pytz==2016.4
django-jet==0.1.4
